--- figures/BCTW_all.ini
@@ -1,20 +1,0 @@
-[global_settings]
-n_instances = 1
-n_prior_included = 1
-n_prior_excluded = 1
-model = nb
-balance_strategy = double
-query_strategy = max
-feature_extraction = tfidf
-
-[balance_param]
-a = 0.7782141491367417
-alpha = 1.39537927110139
-b = 0.9147018292161516
-
-[feature_param]
-ngram_max = 2
-split_ta = 0
-
-[model_param]
-alpha = 2.1292734234216257
