--- figures/README.md
@@ -1,66 +1,0 @@
-# paper-asreview-introduction
-This directory contains scripts to reproduce reproduce the results of Figure 4 for the paper ' ASReview: Open Source Software for Efficient and Transparent Systematic Reviews'.
-
-Figure 4 displays the performance of an Automated Systematic Review model (Naive Bayes + TF-IDF + Certainty + Dynamic Supersampling) over 15 runs on four already labeled datasets: (A) the virus dataset, (B) the hall dataset, (C) PTSD dataset, (4) ACE dataset.
-
-The four datasets are available ... and the simulation output is available on the [Open Science Framework](https://osf.io/2jkd6/).
-
-
-# Install packages
-Make sure you've installed ASReview v0.7.2, the ASReview simulations package and the ASReview visualization package.
-
-```
-pip install asreview
-pip install asreview-simulation
-pip install asreview-visualization
-```
-
-# Simulation
-run the following lines one by one in the bash shell to simulate an Automated Systematic Review on the four datasets
-To reproduce the simulation results available at the OSF, run the following:
-
-#### Figure 4A - Virus dataset
-```
-mpirun -n 2 asreview batch data/virus.csv --config_file BCTW_all.ini --state_file "simoutput/virus/results.h5" -r 15 --server_job
-```
-
-#### Figure 4B - Hall dataset
-```
-mpirun -n 2 asreview batch data/hall.csv --config_file BCTW_all.ini --state_file "simoutput/hall/results.h5" -r 15 --server_job
-```
-#### Figure 4C - PTSD dataset
-```
-mpirun -n 2 asreview batch data/ptsd.csv --config_file BCTW_all.ini --state_file "simoutput/ptsd/results.h5" -r 15 --server_job
-```
-
-#### Figure 4D - ACE dataset
-```
-mpirun -n 2 asreview batch data/ace.csv --config_file BCTW_all.ini --state_file "simoutput/ace/results.h5" -r 15 --server_job
-```
-
----
-
-# Visualization
-The simulation output is used to create Figure 4.
-If you don't have the time or resources to run the simulations yourself, you can also download them from the Open Science Framework page.
-
-#### Figure 4A
-
-```
-python ./scripts/pstd_basic.py
-```
-
-#### Figure 4B
-```
-python ./scripts/pstd_basic.py
-```
-
-#### Figure 4C
-```
-python ./scripts/pstd_basic.py
-```
-
-#### Figure 4D
-```
-python ./scripts/pstd_basic.py
-```
