+++ simulation_study/visualization/README.md
@@ -1,0 +1,24 @@
+# Visualization
+The simulation output is used to create Figure 4.
+If you don't have the time or resources to run the simulations yourself, you can also download them from the Open Science Framework page.
+
+#### Figure 4A
+
+```
+python ./scripts/pstd_basic.py
+```
+
+#### Figure 4B
+```
+python ./scripts/pstd_basic.py
+```
+
+#### Figure 4C
+```
+python ./scripts/pstd_basic.py
+```
+
+#### Figure 4D
+```
+python ./scripts/pstd_basic.py
+```
